% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_data.r
\name{create_new_data}
\alias{create_new_data}
\title{Create new spatial data for interpolation}
\usage{
create_new_data(obj, gen_mode = "chull", npoints = 10000)
}
\arguments{
\item{obj}{A Spatial*DataFrame.}

\item{gen_mode}{character. One of 'rect' (rectangular) and 'chull' (convex hull).}

\item{npoints}{integer. the number of points that will be generated}
}
\value{
A SpatialPointsDataFrame.
}
\description{
Create new spatial data for interpolation
}
