% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_data.r
\name{create_new_data.ST}
\alias{create_new_data.ST}
\title{Generate a new spatiotemporal points for the spatiotemporal prediction and interpolation}
\usage{
create_new_data.ST(obj, form, gen_mode = "chull", npoints = 10000, forward = 6)
}
\arguments{
\item{obj}{a ST*DF object.}

\item{form}{formula.}

\item{gen_mode}{character. One of 'rect' (rectangular) and 'chull' (convex hull).}

\item{npoints}{integer. the number of points that will be generated}

\item{forward}{integer. the time length of the data will generate ahead of the last time point of the input data. If NULL is passed, the spatiotemporal interpolation mode in obj will be conducted.}
}
\value{
A STFDF object.
}
\description{
Generate a new spatiotemporal points for the spatiotemporal prediction and interpolation
}
