% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_data.r
\name{detect_temporal_unit.generic}
\alias{detect_temporal_unit.generic}
\title{Autodetect the temporal unit in a lubridate/POSIXct object}
\usage{
detect_temporal_unit.generic(temporal)
}
\arguments{
\item{temporal}{a lubridate/POSIXct object.}
}
\value{
A character that indicates the temporal unit of the input xts object.
}
\description{
Autodetect the temporal unit in a lubridate/POSIXct object
}
