% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_data.r
\name{detect_temporal_unit.xts}
\alias{detect_temporal_unit.xts}
\title{Autodetect the temporal unit in a xts object}
\usage{
detect_temporal_unit.xts(temporal)
}
\arguments{
\item{temporal}{a xts object.}
}
\value{
A character that indicates the temporal unit of the input xts object.
}
\description{
Autodetect the temporal unit in a xts object
}
