% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal.variogramST.r
\name{marginal.variogramST}
\alias{marginal.variogramST}
\title{Compute the marginal spatial or temporal sample variogram}
\usage{
marginal.variogramST(stv, bound, spatial = TRUE)
}
\arguments{
\item{stv}{A STFDF.}

\item{bound}{numeric. The maximum distance that will be used to compute a spatial variogram.}

\item{spatial}{boolean. if TRUE, the spatial marginal variogram will be obtained, temporal otherwise.}
}
\value{
A gstatVariogram object.
}
\description{
Compute the marginal spatial or temporal sample variogram
}
